型定義をファイル毎に分割しました。
