+++ Sources/YumemiWeather/Area.swift
@@ -1,0 +1,18 @@
+//
+//  Area.swift
+//
+
+public enum Area: String, CaseIterable, Codable {
+    case sapporo = "Sapporo"
+    case sendai = "Sendai"
+    case niigata = "Niigata"
+    case kanazawa = "Kanazawa"
+    case tokyo = "Tokyo"
+    case nagoya = "Nagoya"
+    case osaka = "Osaka"
+    case hiroshima = "Hiroshima"
+    case kochi = "Kochi"
+    case fukuoka = "Fukuoka"
+    case kagoshima = "Kagoshima"
+    case naha = "Naha"
+}
