+++ Sources/YumemiWeather/AreaRequest.swift
@@ -1,0 +1,10 @@
+//
+//  AreaRequest.swift
+//
+
+import Foundation
+
+struct AreaRequest: Decodable {
+    let areas: [String]
+    let date: Date
+}
