+++ Sources/YumemiWeather/AreaResponse.swift
@@ -1,0 +1,8 @@
+//
+//  AreaResponse.swift
+//
+
+struct AreaResponse: Codable {
+    let area: Area
+    let info: Response
+}
