+++ Sources/YumemiWeather/Request.swift
@@ -1,0 +1,10 @@
+//
+//  Request.swift
+//
+
+import Foundation
+
+struct Request: Decodable {
+    let area: String
+    let date: Date
+}
