+++ Sources/YumemiWeather/Response.swift
@@ -1,0 +1,12 @@
+//
+//  File.swift
+//  
+
+import Foundation
+
+struct Response: Codable, Equatable {
+    let weatherCondition: String
+    let maxTemperature: Int
+    let minTemperature: Int
+    let date: Date
+}
