+++ Sources/YumemiWeather/YumemiWeatherError.swift
@@ -1,0 +1,8 @@
+//
+//  YumemiWeatherError.swift
+//
+
+public enum YumemiWeatherError: Error {
+    case invalidParameterError
+    case unknownError
+}
